refactor: Update favicon and manifest file for Emoji Picker
The code changes involve updating the favicon and manifest file in the `public/favicon` directory. This includes adding a name and short name for the site, as well as a description for the Emoji Picker. The manifest file now includes a start URL for the site. These changes improve the appearance and accessibility of the Emoji Picker.